asp pomoyka

+++ Sebezhko/Models/OrdersDetails.cs
@@ -4,7 +4,7 @@
     {
         public int ID { get; set; }
         public int Orders_ID { get; set; }
-        public decimal Product_ID { get; set; }
+        public int Product_ID { get; set; }
 
     }
 }
